% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moidecoder.R
\name{decoder_str}
\alias{decoder_str}
\title{Transforms entries to the desired coding class}
\usage{
decoder_str(y, c_l, r_w, coding, rw_col, multsh)
}
\arguments{
\item{y}{numeric vector; entries in a cell corresponding
to a specific sample and a specific marker.}

\item{c_l}{string; marker label.}

\item{r_w}{numeric; sample ID's row number in the excel
file.}

\item{coding}{string; coding class of the molecular
marker.}

\item{rw_col}{string vector; variable used to switch
between row and column in case of transposed data.
Namely, \code{c("rows ", "row ", "column ", "columns
")}.}

\item{multsh}{string; reports warnings for multiple
worksheet datasets.}
}
\value{
a list of two elements: 1) a vector of STR
  entries in pre-specified coding class. 2) an identifier
  whose value is 1 if a warning takes place.
}
\description{
Transforms the data entries in a cell to a
  pre-specified coding class.
}
\seealso{
For further details see: \code{\link{moimport}},
  \code{\link{moi_marker}}.
}
\keyword{internal}
