% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moianalyses.R
\name{mle}
\alias{mle}
\title{Derives MLE}
\usage{
mle(nnk, nn)
}
\arguments{
\item{nnk}{numeric vector; vector of lineage prevalence
counts.}

\item{nn}{numeric; sample size}
}
\value{
a list with the following elements: 1) log
  likelihood at MLE, MLE of lambda and MLE of psi, 2) MLE
  of lineage frequencies.
}
\description{
Derives MLE
}
\seealso{
For further details see: \code{\link{moimle}}
}
\keyword{internal}
