% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moianalyses.R
\name{mle_fixed}
\alias{mle_fixed}
\title{Derives profile-likelihood MLE of lineage frequencies}
\usage{
mle_fixed(lambda, nnk)
}
\arguments{
\item{lambda}{numeric; MOI parameter}

\item{nnk}{numeric vector; vector of lineage prevalence
counts.}
}
\value{
vector of lineage frequency estimates
}
\description{
Derives profile-likelihood MLE of lineage frequencies
}
\seealso{
For further details see: \code{\link{moimle}}
}
\keyword{internal}
