% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_administrator}
\alias{moi_administrator}
\title{Administrator function}
\usage{
moi_administrator(set_d, s_total, m_total, nummtd, cha, rw_col, nwsh)
}
\arguments{
\item{set_d}{data frame; imported dataset.}

\item{s_total}{string vector; vector of sample IDs.}

\item{m_total}{string vector;vector of marker labels.}

\item{nummtd}{numeric; number of metadata columns plus 2.}

\item{cha}{string vector; vector of punctuation
characters. See \code{\link{moi_prerequisite}}.}

\item{rw_col}{string vector; variable used to switch
between row and column in case of transposed data.
Namely, \code{c("rows ", "row ", "column ", "columns
")}.}

\item{nwsh}{numeric; worksheet number in multiple
worksheet dataset.}
}
\value{
a list of following elements: 1) rows in which
  new samples start, 2) all sample IDs in the worksheet,
  3) number of samples in the worksheet, 3) multiple row
  per sample identifier, 4) another multiple row per
  sample identifier, 5) columns in which new markers
  start, 6) all marker labels in the worksheet, 7) number
  of markers in the worksheet, 8) multiple column per
  marker identifier, 9) another multiple column per
  marker identifier, 10) an identifier whose value is 1
  if a warning takes place.
}
\description{
Administrator function
}
\seealso{
For further details see: \code{\link{moimport}}.
}
\keyword{internal}
