% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_export}
\alias{moi_export}
\title{Exports the dataset in standard format to a new excel
file.}
\usage{
moi_export(export, final)
}
\arguments{
\item{export}{string; the path where the imported data is
stored in standardized format.}

\item{final}{data frame; modified dataset in standard
format.}
}
\value{
An excel file which contains dataset in standard
  format.
}
\description{
This function exports the modified dataset
  in standard format to a new excel file.
}
\seealso{
For further details, please see the following
  functions: \code{moimport}.
}
\keyword{internal}
