% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_labels}
\alias{moi_labels}
\title{Finds sample IDs contained in a dataset}
\usage{
moi_labels(s_total)
}
\arguments{
\item{s_total}{sample IDs.}
}
\value{
a list which the first element is a numeric
  vector specifying the number of rows in which a new
  sample starts. The second element is the name of sample
  IDs.
}
\description{
Each dataset consists of several number of
  samples. Samples are specified by their sample ID which
  are placed in the first column of the excel worksheet.
  The function \code{moi_labels}, finds sample IDs and
  the row in which they start.
}
\seealso{
For further details see: \code{moimport}
  and \code{\link{moi_duplicatefinder}}.
}
\keyword{internal}
