% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_mergemetadata}
\alias{moi_mergemetadata}
\title{Merges metadata}
\usage{
moi_mergemetadata(tempmtd, mtdall, samall, multsh)
}
\arguments{
\item{tempmtd}{matrix; matrix of temporary metadata.}

\item{mtdall}{string vector; vector of all metadata
labels.}

\item{samall}{string vector; vector of all sample IDs in
the worksheet.}

\item{multsh}{string; reports warnings for multiple
worksheet datasets.}
}
\value{
list of following elements: 1) unique metadata
  labels, 2) unique metadata for different samples.
}
\description{
Merges metadata
}
\seealso{
For further details see: \code{moimport}.
}
\keyword{internal}
