% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_metadata}
\alias{moi_metadata}
\title{Checks the metadata entries}
\usage{
moi_metadata(
  metadata,
  mtdlabels,
  nummtd,
  samorder,
  samall,
  lsam,
  nomtdneeded,
  nomerge,
  multsh
)
}
\arguments{
\item{metadata}{matrix; matrix of metadata columns.}

\item{mtdlabels}{string vector; vector of metadata
labels.}

\item{nummtd}{numeric; number of metadata columns plus 2.}

\item{samorder}{numeric vector; rows where samples start.}

\item{samall}{string vector; vector of sample IDs in the
worksheet.}

\item{lsam}{numeric; number of samples in the worksheet.}

\item{nomtdneeded}{numeric vector; samples which need no
metadata.}

\item{multsh}{string; reports warnings for multiple
worksheet datasets.}
}
\value{
a list of following elements: 1) unique metadata
  for different samples, 2) an identifier whose value is
  1 if a warning takes place.
}
\description{
Checks if a sample has unique metadata
  entries.
}
\seealso{
For further details see: \code{moimport}.
}
\keyword{internal}
