% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moianalyses.R
\name{moi_nk}
\alias{moi_nk}
\title{Derives lineage prevalence counts}
\usage{
moi_nk(datmarker, samorder)
}
\arguments{
\item{datmarker}{vector; a column of data corresponding
to a marker from the imported dataset.}

\item{samorder}{numeric vector; row numbers in excel file
where the new samples start.}
}
\value{
a list with two elements: 1) sample size of the
  marker, 2) vector of lineage prevalence counts at the
  marker.
}
\description{
Derives lineage prevalence counts
}
\seealso{
For further details see: \code{\link{moimle}}
}
\keyword{internal}
