% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_prerequisite}
\alias{moi_prerequisite}
\title{Contains different character vectors needed for importing
dataset}
\usage{
moi_prerequisite()
}
\description{
Contains different character vectors needed for importing
dataset
}
\keyword{internal}
