% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_separator}
\alias{moi_separator}
\title{Finds the most frequent sparator}
\usage{
moi_separator(set_d, nummtd, cha, rw_col, nwsh)
}
\arguments{
\item{set_d}{data frame; imported dataset.}

\item{nummtd}{numeric; number of metadata coumns plus 2.}

\item{cha}{string vector; vector of punctuation
characters. See \code{\link{moi_prerequisite}}.}

\item{rw_col}{string vector; variable used to switch
between row and column in case of transposed data.
Namely, \code{c("rows ", "row ", "column ", "columns
")}.}

\item{nwsh}{numeric; worksheet number in multiple
worksheet dataset.}
}
\value{
a warning is generated reporting inconsistencies
  in usage of separator.
}
\description{
This function is activated when the data is
  of 'One row per sample, one column per marker' format.
  In such a dataset, the user enters multiple information
  corresponding to a sample on a marker in one cell and
  separates the items with a symbol (punctuation
  character). It is expected that user be consistent with
  usage of symbol. Otherwise this function addresses the
  inconsistencies with a warning.
}
\seealso{
For further details see: \code{\link{moimport}}.
}
\keyword{internal}
