% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{withWarnings}
\alias{withWarnings}
\title{Captures all the warnings}
\usage{
withWarnings(expr)
}
\arguments{
\item{expr}{expression to be evaluated.}
}
\value{
list with two elements. First is the value of the
  function. Second is the generated functions by the
  expression.
}
\description{
Captures all the warnings
}
\seealso{
For further details, please see the following
  functions: \code{moimport}.
}
\keyword{internal}
