% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moibasic.R
\name{moi_empty}
\alias{moi_empty}
\title{Reports and deletes empty rows/columns}
\usage{
moi_empty(set_d, nummtd)
}
\arguments{
\item{set_d}{data frame; imported dataset.}

\item{nummtd}{numeric; number of metadata coumns plus 2.}
}
\value{
A list of three elements: 1) list of empty rows,
  2) list of empty columns, 3) contains the dataset with
  deleted empty rows/columns.
}
\description{
Reports and deletes empty rows/columns.
}
\seealso{
For further details see: \code{\link{moimport}}.
}
\keyword{internal}
