% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moimerge.R
\name{moimerge}
\alias{moimerge}
\title{Merges two molecular datasets}
\usage{
moimerge(file1, file2, nummtd1, nummtd2, keepmtd = F, export = NULL)
}
\arguments{
\item{file1}{string; specifying the path of the first
dataset.}

\item{file2}{string; specifying the path of the second
dataset.}

\item{nummtd1}{numeric; number of metadata columns (see
\code{\link{moimport}}) in the first file (default as
\code{0}).}

\item{nummtd2}{numeric; number of metadata columns (see
\code{\link{moimport}}) in the second file (default as
\code{0}).}

\item{keepmtd}{logical; determining whether metadata
(e.g., date) should be retained (default as
\code{TRUE}).}

\item{export}{string; the path where the data is stored.}
}
\value{
The output is a dataset in standard format which
  constituts of an assembly of the input datasets.
}
\description{
The function is designed to merge two
  datasets from separate \file{.xlsx}-files. The data in
  each \file{.xlsx}-files is placed in the first
  worksheet.
}
\details{
The two datasets should be already in standard
  format (see (see \code{\link{moimport}})). The datasets
  are placed in the first worksheet of the two different
  \file{.xlsx}-files. Notice that marker labels (=column
  labels) need to be unique.
}
\section{Warnings}{
 Warnings are generated if potential
  inconsistencies are detected. E.g.if the same sample
  occurs in both datasets and have contradicting metadata
  entries.
}

\examples{
infile1 <- system.file("/extdata/", "testDatamerge1.xlsx", package = "tempMLMOI")
infile2 <- system.file("/extdata/", "testDatamerge2.xlsx", package = "tempMLMOI")
moimerge(infile1, nummtd1 = 1, infile2, nummtd2 = 2, keepmtd = TRUE)

}
\seealso{
To import and transform data into standard
  format, please see the function \code{\link{moimport}}.
}
